ranking/ch3/train_reranker/requirements.txt
requests
openai
pyvespa
python-dotenv
numpy
pandas
lightgbm
plotly
nbformat
scikit-learn
ipynbname